# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Programmatic access to the Weblens server",
        "title": "Weblens API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/",
    "paths": {
        "/files": {
            "delete": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Delete Files \"permanently\"",
                "operationId": "DeleteFiles",
                "parameters": [
                    {
                        "description": "Delete files request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/FilesListParams"
                        }
                    },
                    {
                        "type": "boolean",
                        "description": "Delete files even if they are not in the trash",
                        "name": "ignoreTrash",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Move a list of files to a new parent folder",
                "operationId": "MoveFiles",
                "parameters": [
                    {
                        "description": "Move files request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MoveFilesParams"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Share Id",
                        "name": "shareId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/autocomplete": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Get path completion suggestions",
                "operationId": "AutocompletePath",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search path",
                        "name": "searchPath",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Path info",
                        "schema": {
                            "$ref": "#/definitions/FolderInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/search": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Search for files by filename",
                "operationId": "SearchByFilename",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filename to search for",
                        "name": "search",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The folder to search in, defaults to the user's home folder",
                        "name": "baseFolderId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File Info",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/FileInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/shared": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Get files shared with the logged in user",
                "operationId": "GetSharedFiles",
                "responses": {
                    "200": {
                        "description": "All the top-level files shared with the user",
                        "schema": {
                            "$ref": "#/definitions/FolderInfo"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/structsore": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "structsore files from some time in the past",
                "operationId": "RestoreFiles",
                "parameters": [
                    {
                        "description": "RestoreFiles files request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RestoreFilesBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "structsore files info",
                        "schema": {
                            "$ref": "#/definitions/RestoreFilesInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/untrash": {
            "patch": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Move a list of files out of the trash, structsoring them to where they were before",
                "operationId": "UnTrashFiles",
                "parameters": [
                    {
                        "description": "Un-trash files request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/FilesListParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/{fileId}": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Get information about a file",
                "operationId": "GetFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File Id",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Share Id",
                        "name": "shareId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File Info",
                        "schema": {
                            "$ref": "#/definitions/FileInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Update a File",
                "operationId": "UpdateFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File Id",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Share Id",
                        "name": "shareId",
                        "in": "query"
                    },
                    {
                        "description": "Update file request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateFileParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/{fileId}/download": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Download a file",
                "operationId": "DownloadFile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File Id",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Share Id",
                        "name": "shareId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "File format conversion",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": false,
                        "description": "Is this a takeout file",
                        "name": "isTakeout",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Error Info",
                        "schema": {
                            "$ref": "#/definitions/WeblensErrorInfo"
                        }
                    }
                }
            }
        },
        "/files/{fileId}/history": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "tags": [
                    "Folder"
                ],
                "summary": "Get actions of a folder at a given time",
                "operationId": "GetFolderHistory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File Id",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Past timestamp to view the folder at, in ms since epoch",
                        "name": "timestamp",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File actions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/FileActionInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/files/{fileId}/stats": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Get the statistics of a file",
                "operationId": "GetFileStats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File Id",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "501": {
                        "description": "Not Implemented"
                    }
                }
            }
        },
        "/files/{fileId}/text": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Get the text of a text file",
                "operationId": "GetFileText",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File Id",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Share Id",
                        "name": "shareId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/folder": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folder"
                ],
                "summary": "Create a new folder",
                "operationId": "CreateFolder",
                "parameters": [
                    {
                        "description": "New folder body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateFolderBody"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Share Id",
                        "name": "shareId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File Info",
                        "schema": {
                            "$ref": "#/definitions/FileInfo"
                        }
                    }
                }
            }
        },
...
        "/folder/{folderId}": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folder"
                ],
                "summary": "Get a folder",
                "operationId": "GetFolder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder Id",
                        "name": "folderId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Share Id",
                        "name": "shareId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Past timestamp to view the folder at, in ms since epoch",
                        "name": "timestamp",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder Info",
                        "schema": {
                            "$ref": "#/definitions/FolderInfo"
                        }
                    }
                }
            }
        },
        "/folder/{folderId}/cover": {
            "patch": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "tags": [
                    "Folder"
                ],
                "summary": "Set the cover image of a folder",
                "operationId": "SetFolderCover",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder Id",
                        "name": "folderId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Media Id",
                        "name": "mediaId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/keys": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKeys"
                ],
                "summary": "Get all api keys",
                "operationId": "GetApiKeys",
                "responses": {
                    "200": {
                        "description": "Tokens",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TokenInfo"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKeys"
                ],
                "summary": "Create a new api key",
                "operationId": "CreateApiKey",
                "parameters": [
                    {
                        "description": "The new token params",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ApiKeyParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The new token",
                        "schema": {
                            "$ref": "#/definitions/TokenInfo"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/keys/{tokenId}": {
            "delete": {
                "security": [
                    {
                        "SessionAuth": [
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApiKeys"
                ],
                "summary": "Delete an api key",
                "operationId": "DeleteApiKey",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Api key id",
                        "name": "tokenId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/users/logout": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Logout User",
                "operationId": "LogoutUser",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "FilesListParams": {
            "type": "object",
            "properties": {
                "fileIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "MoveFilesParams": {
            "type": "object",
            "properties": {
                "fileIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "newParentId": {
                    "type": "string"
                }
            }
        },
...
        "RestoreFilesInfo": {
            "type": "object",
            "properties": {
                "newParentId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "scopes": {
                "admin": "Grants read and write access to privileged data"
            }
        },
        "SessionAuth": {
            "type": "apiKey",
            "name": "weblens-session-token",
            "in": "cookie"
        }
    }
}